Added unit tests and fixes to doc table

+++ src/tests/test_util.h
@@ -28,10 +28,10 @@
 
 #define ASSERT_STRING_EQ(s1, s2) ASSERT(!strcmp(s1, s2));
 
-#define ASSERT_EQUAL_INT(x, y, ...)                       \
-  if (x != y) {                                           \
-    fprintf(stderr, "%d != %d: " __VA_ARGS__ "\n", x, y); \
-    return -1;                                            \
+#define ASSERT_EQUAL_INT(x, y, ...)                                 \
+  if (x != y) {                                                     \
+    fprintf(stderr, "%d != %d: " __VA_ARGS__ "\n", (int)x, (int)y); \
+    return -1;                                                      \
   }
 
 #define FAIL(fmt, ...)                                                            \
